+++ jwebform/src/main/java/jwebform/field/helper/CommonSelects.java
@@ -1,33 +1,26 @@
 package jwebform.field.helper;
 
-import jwebform.field.SelectType;
-
 import java.util.ArrayList;
 import java.util.List;
+import jwebform.field.SelectType;
 
 /**
  * Common convenience setups for select fields.
  * <p>
- * Build-in:
- * * Months (january ... december)
- * * Days (1..31)
- * * Years (start ... end)
- * * Gender
+ * Build-in: * Months (january ... december) * Days (1..31) * Years (start ... end) * Gender
  * <p>
  * RFE: Add Days (monday...sunday),
  */
 public class CommonSelects {
 
-  private CommonSelects() {
-  }
+  private CommonSelects() {}
 
   public static CommonSelects build() {
     return new CommonSelects();
   }
 
-  private static final String[] months =
-    {"january", "february", "march", "april", "may", "june", "july", "august", "september",
-      "october", "november", "december"};
+  private static final String[] months = {"january", "february", "march", "april", "may", "june",
+      "july", "august", "september", "october", "november", "december"};
 
 
   public List<SelectType.SelectInputEntry> buildDays() {
@@ -40,19 +33,27 @@ public class CommonSelects {
     monthsList.add(new SelectType.SelectInputEntry("", "jwebform.select.month"));
     for (int i = 1; i <= 12; i++) {
       monthsList.add(
-        new SelectType.SelectInputEntry(Integer.toString(i), "jwebform.select." + months[i - 1]));
+          new SelectType.SelectInputEntry(Integer.toString(i), "jwebform.select." + months[i - 1]));
     }
     return monthsList;
   }
 
+  /**
+   * will always return a list of years from startYear to endYear. Example:
+   * 
+   * startyear: 2010, endyear: 2012: => 2010,2011,2012 startyear: 2012, endyear: 2010: =>
+   * 2012,2011,2010
+   * 
+   * @param startYear
+   * @param endYear
+   * @return
+   */
   public List<SelectType.SelectInputEntry> getYears(int startYear, int endYear) {
     if (startYear < endYear) {
-      int tmp = startYear;
-      startYear = endYear;
-      endYear = tmp;
+      return buildMap(startYear, endYear, "jwebform.select.year");
+    } else {
+      return builReverseMap(startYear, endYear, "jwebform.select.year");
     }
-    List<SelectType.SelectInputEntry> yearMap =  builReverseMap(startYear, endYear, "jwebform.select.year");
-    return yearMap;
   }